+++ crates/settings_ui/src/settings_ui.rs
@@ -515,6 +515,7 @@ fn init_renderers(cx: &mut App) {
         .add_basic_renderer::<settings::ShellDiscriminants>(render_dropdown)
         .add_basic_renderer::<settings::RelativeLineNumbers>(render_dropdown)
         .add_basic_renderer::<settings::WindowDecorations>(render_dropdown)
+        .add_basic_renderer::<settings::OptionAsAlt>(render_dropdown)
         .add_basic_renderer::<settings::FontSize>(render_editable_number_field)
         .add_basic_renderer::<settings::Decoration>(render_editable_number_field)
         .add_basic_renderer::<settings::SupertabFallback>(render_dropdown)