+++ crates/workspace/src/workspace_settings.rs
@@ -34,6 +34,7 @@ pub struct WorkspaceSettings {
     pub use_system_window_tabs: bool,
     pub zoomed_padding: bool,
     pub window_decorations: settings::WindowDecorations,
+    pub option_as_alt: settings::OptionAsAlt,
 }
 
 #[derive(Copy, Clone, PartialEq, Debug, Default)]
@@ -109,6 +110,7 @@ impl Settings for WorkspaceSettings {
             use_system_window_tabs: workspace.use_system_window_tabs.unwrap(),
             zoomed_padding: workspace.zoomed_padding.unwrap(),
             window_decorations: workspace.window_decorations.unwrap(),
+            option_as_alt: workspace.option_as_alt.unwrap(),
         }
     }
 }